IR: split out ReScript parser
